10th Commit: Finished with the Form Handling
Finished the validation of the required form values and also added a basic encryption to prevent hacking through XSS. Added additional functions on successful registration and in the case of an error. 
Disclaimer:  Form values disappear on submit despite error message popping so users would have to start from scratch again, so I might have to add a feature that saves the values to the local storage and then pull them from there on refresh or page reload. Other methods will be researched.